during bundle generate, ask the user where they want to output the bundle
