+++ pkg/bundle/templates/terraform/.github/workflows/publish.yaml
