+++ pkg/bundle/templates/terraform/.gitignore
