+++ pkg/bundle/templates/terraform/.pre-commit-config.yaml
