+++ pkg/bundle/templates/terraform/operator.mdx
