+++ pkg/bundle/templates/terraform/schema.stories.js
