+++ pkg/bundle/templates/terraform/src/_artifacts.tf
