+++ pkg/bundle/templates/terraform/src/_providers.tf
