+++ pkg/bundle/templates/terraform/src/main.tf
