--- pkg/generator/templates/terraform/src/dev.connections.tfvars.json
@@ -1,3 +1,0 @@
-{
-  "__comment": "Place development connections values here in JSON form. This file is ignored so secrets won't be checked in"
-}
