--- pkg/generator/templates/terraform/src/dev.params.tfvars.json
@@ -1,11 +1,0 @@
-{
-  "md_metadata": {
-    "name_prefix": "local-dev-replaceme-0000",
-    "default_tags": {
-      "md-project": "local",
-      "md-target": "dev",
-      "md-manifest": "replaceme",
-      "md-package": "local-dev-replaceme-0000"
-    }
-  }
-}
